Moved middleware outside of the root folder to fix build
